pyproject.toml
[project]
    authors = [
        { name = "Joaquín Zúñiga", email = "jp.zuniga.dev@gmail.com" },
    ]

    dependencies = [
        "dj-database-url>=3.1.0",
        "django-cors-headers>=4.9.0",
        "django>=6.0",
        "djangorestframework-simplejwt>=5.5.1",
        "djangorestframework>=3.16.1",
        "gunicorn>=23.0.0",
        "psycopg[binary]>=3.3.2",
        "python-dotenv>=1.2.1",
        "whitenoise>=6.11.0",
    ]

    description     = "Barebones Django API."
    name            = "api"
    readme          = "README.md"
    requires-python = ">=3.13.9"
    version         = "0.1.0"

[dependency-groups]
    dev = [
        "factory-boy>=3.3.3",
        "faker>=40.1.2",
        "freezegun>=1.5.5",
        "pytest-cov>=7.0.0",
        "pytest-django>=4.11.1",
        "pytest>=9.0.2",
        "ruff>=0.14.13",
    ]

[tool.coverage.run]
    branch = true
    omit = [
        "**/__init__.py",
        "**/admin/*",
        "**/apps.py",
        "**/consts.py",
        "**/migrations/*",
        "**/tests/*",
    ]

    source = ["api"]

[tool.coverage.report]
    omit = [
        "**/__init__.py",
        "**/admin/*",
        "**/consts.py",
        "**/apps.py",
        "**/migrations/*",
        "**/tests/*",
    ]

    show_missing = true

[tool.pytest.ini_options]
    DJANGO_SETTINGS_MODULE = "core.settings"
    addopts                = "--cov=api --cov-config=pyproject.toml --cov-report=term-missing --reuse-db -q"
    console_output_style   = "count"
    markers                = ["unit"]
    pythonpath             = "src"

[tool.ruff.lint]
    allowed-confusables = ["−"]
    ignore = [
        "COM812",
        "D100",
        "D101",
        "D102",
        "D103",
        "D104",
        "D105",
        "D106",
        "D107",
        "D200",
        "D202",
        "D203",
        "D212",
    ]

    select = ["ALL"]

    [tool.ruff.lint.isort]
        case-sensitive         = true
        combine-as-imports     = true
        force-wrap-aliases     = true
        order-by-type          = true
        relative-imports-order = "closest-to-furthest"
        required-imports       = ["from __future__ import annotations"]

    [tool.ruff.lint.per-file-ignores]
        "**/apps.py" = ["F401", "PLC0415"]
        "**/manage.py" = ["PLC0415"]
        "**/migrations/*" = [
            "ANN001",
            "ARG001",
            "D100",
            "D101",
            "D103",
            "D104",
            "E501",
            "RUF012",
        ]

        "**/settings.py" = ["E501"]
        "**/tests/*" = [
            "D100",
            "D101",
            "D102",
            "D103",
            "D104",
            "D106",
            "PT013",
            "S101",
            "S106",
        ]
